CLEANED-UP: moved Directory into its separate FileState file
